Decouple blockchain analysis from transaction processor.
